refactor(locale): share interpolate helper, guard lookupEn, fix onMissing example
Addresses review feedback on the English-fallback feature:

- Extract the placeholder-interpolation grammar into a single
  useLocale/interpolate.ts shared by V0LocaleAdapter and the English
  fallback lookup, removing the near-verbatim copy in index.ts (and the
  comment that claimed to "reuse" the adapter while reimplementing it).
- Replace raw typeof/=== null/=== undefined checks in lookupEn with the
  #v0/utilities guards (isObject/isString), matching repo convention.
- Fix the onMissing JSDoc example: en is nested, so en[key] never resolved
  a dotted key; show a correctly-keyed flat dictionary instead.

Behavior unchanged — useLocale + Avatar/Breadcrumbs/NumberField/Pagination
specs all pass.
